+++ lib/td_dd/canada/data_structure_abilities.ex
@@ -10,16 +10,23 @@ defmodule TdDd.Canada.DataStructureAbilities do
       Permissions.authorized?(user, :create_data_structure, domain_id)
   end
 
-  def can?(%User{} = user, :delete_data_structure, %DataStructure{domain_id: domain_id}) do
-    Permissions.authorized?(user, :delete_data_structure, domain_id)
+  def can?(%User{} = user, :manage_confidential_structures, %DataStructure{domain_id: domain_id}) do
+    Permissions.authorized?(user, :manage_confidential_structures, domain_id)
   end
 
-  def can?(%User{} = user, :update_data_structure, %DataStructure{domain_id: domain_id}) do
-    Permissions.authorized?(user, :update_data_structure, domain_id)
+  def can?(%User{} = user, :delete_data_structure, %DataStructure{domain_id: domain_id, confidential: confidential}) do
+    Permissions.authorized?(user, :delete_data_structure, domain_id) &&
+    (!confidential || Permissions.authorized?(user, :manage_confidential_structures, domain_id))
   end
 
-  def can?(%User{} = user, :view_data_structure, %DataStructure{domain_id: domain_id}) do
-    Permissions.authorized?(user, :view_data_structure, domain_id)
+  def can?(%User{} = user, :update_data_structure, %DataStructure{domain_id: domain_id, confidential: confidential}) do
+    Permissions.authorized?(user, :update_data_structure, domain_id) &&
+    (!confidential || Permissions.authorized?(user, :manage_confidential_structures, domain_id))
+  end
+
+  def can?(%User{} = user, :view_data_structure, %DataStructure{domain_id: domain_id, confidential: confidential}) do
+    Permissions.authorized?(user, :view_data_structure, domain_id) &&
+    (!confidential || Permissions.authorized?(user, :manage_confidential_structures, domain_id))
   end
 
   def can?(%User{} = user, :view_data_structure, domain_id) do