+++ test/td_dd_web/controllers/data_structure_controller_test.exs
@@ -3,7 +3,10 @@ defmodule TdDdWeb.DataStructureControllerTest do
   import TdDdWeb.Authentication, only: :functions
   use PhoenixSwagger.SchemaTest, "priv/static/swagger.json"
 
+  alias TdDd.DataStructures
   alias TdDd.DataStructures.DataStructure
+  alias TdDd.MockTaxonomyCache
+  alias TdDd.Permissions.MockPermissionResolver
   alias TdDdWeb.ApiServices.MockTdAuditService
   alias TdDdWeb.ApiServices.MockTdAuthService
 
@@ -43,6 +46,8 @@ defmodule TdDdWeb.DataStructureControllerTest do
   setup_all do
     start_supervised(MockTdAuthService)
     start_supervised(MockTdAuditService)
+    start_supervised(MockPermissionResolver)
+    start_supervised(MockTaxonomyCache)
     start_supervised(@df_cache)
     :ok
   end
@@ -234,9 +239,119 @@ defmodule TdDdWeb.DataStructureControllerTest do
     end
   end
 
+  describe "data_structure confidentiality" do
+    setup [:create_data_structure]
+
+    @tag authenticated_user: @admin_user_name
+    test "updates data_structure confidentiality", %{
+      conn: conn,
+      data_structure: %DataStructure{id: id} = data_structure
+    } do
+      assert Map.get(data_structure, :confidential) == false
+      conn =
+        put(
+          conn,
+          data_structure_path(conn, :update, data_structure),
+          data_structure: %{confidential: true}
+        )
+
+      assert %{"id" => ^id} = json_response(conn, 200)["data"]
+      conn = recycle_and_put_headers(conn)
+
+      conn = get(conn, data_structure_path(conn, :show, id))
+      json_response_data = json_response(conn, 200)["data"]
+
+      assert json_response_data["id"] == id
+      assert json_response_data["confidential"] == true
+    end
+
+    @tag authenticated_no_admin_user: "user"
+    test "user with permission can update confidential data_structure", %{conn: conn, user: %{id: user_id}} do
+      role_name = "confidential_editor"
+      confidential = true
+      data_structure = create_data_structure_and_permissions(user_id, role_name, confidential)
+      %{id: id} = data_structure
+
+      conn =
+        put(
+          conn,
+          data_structure_path(conn, :update, data_structure),
+          data_structure: %{description: "edited desc"}
+        )
+
+      assert %{"id" => ^id} = json_response(conn, 200)["data"]
+      conn = recycle_and_put_headers(conn)
+
+      conn = get(conn, data_structure_path(conn, :show, id))
+      json_response_data = json_response(conn, 200)["data"]
+
+      assert json_response_data["id"] == id
+      assert json_response_data["description"] == "edited desc"
+    end
+
+    @tag authenticated_no_admin_user: "user_without_permission"
+    test "user without confidential permission cannot update confidential data_structure", %{conn: conn, user: %{id: user_id}} do
+      role_name = "editor"
+      confidential = true
+      data_structure = create_data_structure_and_permissions(user_id, role_name, confidential)
+      %{id: id} = data_structure
+
+      conn =
+        put(
+          conn,
+          data_structure_path(conn, :update, data_structure),
+          data_structure: %{description: "edited desc"}
+        )
+
+      assert json_response(conn, 403)
+      new_data_structure = DataStructures.get_data_structure!(id)
+      assert Map.get(new_data_structure, :description) == "some description"
+    end
+
+    @tag authenticated_no_admin_user: "user_without_confidential"
+    test "user without confidential permission cannot update confidentiality of data_structure", %{conn: conn, user: %{id: user_id}} do
+      role_name = "editor"
+      confidential = false
+      data_structure = create_data_structure_and_permissions(user_id, role_name, confidential)
+      %{id: id} = data_structure
+
+      conn =
+        put(
+          conn,
+          data_structure_path(conn, :update, data_structure),
+          data_structure: %{confidential: true}
+        )
+
+      assert json_response(conn, 200)
+      new_data_structure = DataStructures.get_data_structure!(id)
+      assert Map.get(new_data_structure, :confidential) == false
+    end
+  end
+
   defp create_data_structure(_) do
     data_structure = insert(:data_structure)
-    data_structure_version = insert(:data_structure_version)
+    data_structure_version = insert(:data_structure_version, data_structure_id: data_structure.id)
     {:ok, data_structure: data_structure, data_structure_version: data_structure_version}
+  end
+
+  defp create_data_structure_and_permissions(user_id, role_name, confidential) do
+    domain_name = "domain_name"
+    domain_id = 1
+    MockTaxonomyCache.create_domain(%{name: domain_name, id: domain_id})
+    MockPermissionResolver.create_acl_entry(%{
+      principal_id: user_id,
+      principal_type: "user",
+      resource_id: domain_id,
+      resource_type: "domain",
+      role_name: role_name
+    })
+
+    data_structure = insert(:data_structure,
+      confidential: confidential,
+      name: "confidential",
+      ou: domain_name,
+      domain_id: domain_id)
+    insert(:data_structure_version, data_structure_id: data_structure.id)
+    data_structure
   end
 end